chore(extension): create new extension file to place all extension class
