+++ FearExperience/View/Screen/WarningView.swift
@@ -8,34 +8,6 @@
 
 import UIKit
 
-extension UIButton {
-    
-    func startAnimatingPressActions() {
-        addTarget(self, action: #selector(animateDown), for: [.touchDown, .touchDragEnter])
-        addTarget(self, action: #selector(animateUp), for: [.touchDragExit, .touchCancel, .touchUpInside, .touchUpOutside])
-    }
-    
-    @objc private func animateDown(sender: UIButton) {
-        animate(sender, transform: CGAffineTransform.identity.scaledBy(x: 0.95, y: 0.95))
-    }
-    
-    @objc private func animateUp(sender: UIButton) {
-        animate(sender, transform: .identity)
-    }
-    
-    private func animate(_ button: UIButton, transform: CGAffineTransform) {
-        UIView.animate(withDuration: 0.4,
-                       delay: 0,
-                       usingSpringWithDamping: 0.5,
-                       initialSpringVelocity: 3,
-                       options: [.curveEaseInOut],
-                       animations: {
-                        button.transform = transform
-            }, completion: nil)
-    }
-    
-}
-
 class WarningView: BaseView {
     
     let showWarning: UIImageView = {
